refactor: Split filter component into responsive
